Version 1.3.7
NF: Added new option g:sqlutil_use_tbl_alias, which is on by default.
    When creating a column list for a table, it will add a table
    alias to each column.  The table alias is calculated based on
    rules.  This feature has 3 settings:
       n - no alias
       d - default calculated alias
       a - ask the user using default alias
NF: Updated g:sqlutil_col_list_terminators to handle some additional
    cases.
BF: In some cases -@- could be left in the query.  This seems to be
    a bug in Vim, where the marks could move.  Unsure why, but put
    a workaround in.
BF: When g:sqlutil_align_comma=1, and the lines did not begin with
    commas, the formatting was incorrect.
BF: When searching through buffers (SQLU_CreateColumnList), the
    alternate buffer could be changed.