Version 1.3.8
NF: Enhanced the feature to change the case of keywords, to include the optional join words, left, right, full, inner, outer.  These words were missed prior to this change and only the JOIN word was capitalized.
CH: Changed the default for sqlutil_use_tbl_alias = 'a' (ask) from being on at all times.  This can be overridden in your vimrc.