Version 1.4.0
BF: If a statement has parenthesis inside of strings, the formatter was not ignoring them.  This could result in the formatter reporting there were mismatched parenthesis.
    Example: select 'string' + '(' from dummy
BF: If a keyword was not followed by a space, it was not placed on a newline.
    Example:  SELECT * FROM(T1)
BF: Nested function calls could potentially lead to an endless loop.